Automation: Cancel automation if the screen becomes unavailable
Operations that use the accessibility service may obstruct the user.
In some cases the lockscreen prevented the user from canceling.
So #685 disabled the overlay if the screen is unavailable, e.g. the lockscreen is active.

This change goes a step further.
The overlay is still hidden if the screen becomes unavailable, but we will also throw a special Exception that should propagate up the stack and cancel the whole operation.

So to stop the automation, turning off the screen is enough.
The automation itself keeps the screen from timing-out, so this should only happen if the user manually presses the power button.

Also see #692.

Closes #686

A "pause" option would be inappropriate here, screen becoming unavailable is an exceptional state.