Root/Shizuku: Fix SD Maid crashing on when activating Root or Shizuku
As guessed, was an R8 issue.

Before the AGP update (i.e. on 8.0.2) the DESCRIPTOR field was implicitly kept.
After moving to AGP 8.1.2 and fixing the reflection access (#705) works, but R8 will remove the field.

So this PR adds rules to keep the `DESCRIPTOR` field for Root and Shizuku.

Fixes #721
Fixes #722