SystemCleaner: Prevent early abort due to reading inaccessible paths
SD Maid should skip `Android/data`, `Android/obb` and `Android/media` when scanning the `SDCARD` area as they will be scanned in an additional step if available.
This skipping did not work correctly because it was based on `currentAreas` which would not contain, e.g. `PUBLIC_DATA` aka `Android/data` was not accessible and not setup as detected area.