IO: Fix WriteException for already deleted files in NORMAL mode
`canNormalWrite` also returns `false` if the file is deleted, then we aborted early with an exception.
Instead we want to run into the `exists()` check later on and not throw an exception if the file no longer exists.
This worked for ROOT and ADB, but not NORMAL due to the extra writable check