IO: Improve scan performance of nested directories
Perform path tree walking at lower access levels (i.e. SAF & Local(Normal/ADB/Root)).
This reduces overhead and in allows for more optimization.
In the case of ADB/Shizuku and Root this reduces the back and forth between processes and through the IPC buffer:
Instead of multiple and repeated `lookupFiles` between SD Maid and the Shizuku/Root process, we only have one call and return the list of of all nested files.
The downside is that we can't use filters with anonymous functions anymore as we can't pass those through the IPC buffer.