Setup: Gracefully handle QUERY_ALL_PACKAGES permission missing on some devices
Previous we always assumed that QUERY_ALL_PACKAGES was granted (which it should, by default),
but Samsung seems to be doing "Just Samsung things :tm:" and this may not be the case on some of Samsungs Android 14 ROMs.
So let's treat it as potentially optional and catch the error, don't crash, and guide the user to the setup.

Closes #1015