AppCleaner: Improve ACS based cache clearing resilience by detecting storage entries "on-the-fly" via app size values
Most "Storage" entries in system details have a summary row that says something like "604 MB in internal storage"
While we don't know the translation of "in internal storage", we know the app size, so we can look for "604 MB".
There is a slight chance for false-positives if the data-usage of the app is the same as the app size, but that's pretty rare.
If that happens SD Maid just keeps retrying as usual.

With 218 apps this had a 99% success rate.
There was one app ("Digital Wellbeing", a system app), that showed the wrong app size (36,06 MB) on first load.
Only when SD Maid retried and reloaded the settings details, the size displayed correct (35,96 MB).
No idea why the size was displayed incorrectly the first time...

When clearing caches via ACS we now have 3 tier process:
1: Try to get the storage entry label via string resource ID
2: Try hard coded labels by language code
3: Try to find the summary row by looking up the app size

This should make it a lot easier to automatically support ROMs after updates, where the texts change.