Dashboard: Fix One-Tap-Mode sometimes not executing deletion task
One-Tap-Mode was tied to the viewmodel lifecycle and a combination of the default scan-task and delete-task, if the viewmodel gets collected (e.g. due to ACS opening system settings over SD Maids main UI, than the coroutine that would have launched the follow up deletion task, is cancelled.
The scan task still continues because after a task is launched, it is tied to the global app scope.

So let's fix this by just creating a special One-Click-Task that will internally perform scan + delete.