Setup: Don't hang if setup checks take longer, show a loading state instead :)
Previously setup modules could only emit their state or nothing.
In some cases this lead to the dashboard hanging, because SD Maid was endlessly waiting for the state from the ShizukuSetupModule.
Now every setup module can emit a `Loading` state and then later a `Current` state (does not have to be final).
This allows us to show "loading states" for each module.

The dashboard will now load faster because we are not waiting for any setup states.
Should loading any setup state take longer than 3 seconds, then we will show a setup card with a loading indicator.
If there are loading states, but also incomplete setup steps, then the setup card will directly be shown.
If you enter the setup screen incomplete items will be shown with loading cards in the list.

Closes #1116