Implement release party screen
When SD Maid goes from <1.0.0 to >= 1.0.0 we display a little release party screen.
The screen indicates that non-prereleases are available and helps the user switch to the release track.
For FOSS builds this means internally using the PROD instead of BETA channel for SD Maids updater.
For GPLAY builds it opens the testing opt-out page: https://play.google.com/apps/testing/eu.darken.sdmse

We also save whether the user has had any pre 1.0.0 versions installed to mark them as early adopter.
Later on we use this flag to ask for a re-review of SD Maid as any rating given during testing is not publicly displayed.