IO: Fix WriteException when trying to delete files that don't exist
This usually happens if temporary files are deleted by their owner (or the system) before SD Maid can act on them.
We want to silently treat this as success.
This failed loudly though because `canWrite` returns false on a file that does not exist, so SD Maid thought no matching access mode was available.
A simple `exists()` check here is insufficient as we can't use that to differentiate between a file that does not exist, and us lacking the permission to determine if it exists.
This commit adds a workaround: We check the parent folder, if we can use `listFiles` on it, then this is unlikely to be a permission issue and the file likely does really not exist.