Automation: Better return behavior after accessibility operations
Instead of back-button presses (i.e. `GLOBAL_ACTION_BACK`) that don't work reliably on Android 14 anymore, we use intent navigation to return to SD Maid at the end.
So far, this proved reliable on API29, API31 and API34.
Opening the system settings activity was quick, and returning to SD Maid was quick too. No extra activity remained in the task-switcher.