AppCleaner: Fix accessibility based cache deletion failing on large apps
On large apps, the system may take a while to load their stats.
This is first retrieved when the system gets the data to display it on screen.
The on-the-fly labeler also retrieves the data to compare against though.
Because the labeler does it on-demand this, but each time.
Requesting the size repeatedly, if the system is already working on it, can delay the process.

This PR changes the labeler, so the size is only retrieved once per app.

Fixes #1304