AppCleaner: Fix scan aborting on read errors with root- or shizuku-level access
While it was a nice idea to reconstruct the actual exceptions, this falls short for exception that don't have the default `String` constructor, e.g. `ReadException` or any `PathException`.
Instead we throw a new `WrappedIPCException` that should be easily searchable, and still offer the original message and a somewhat accurate stacktrace.