AppCleaner: Improve dynamic storage entry label detection
Improve `OnTheFlyLabler` by also manually calculating the formatted size value.
On a VIVO device with Android 14 (`vivo/PD2366/PD2366:14/UP1A.231005.007_MOD1/compiler07161632:user/release-keys`), the `formatFileSize` was the same as the `formatShortFileSize` despite the system showing a non "short" value.

The crawler showed that the storage entry looks like this

```
2024-07-30T08:08:44.858Z  V/SDMSE:Automation:Crawler: crawl():------------12: className=android.widget.TextView, text='Storage', isClickable=false, isEnabled=true, viewIdResourceName=com.android.settings:id/title, pkgName=com.android.settings, identity=6377116
2024-07-30T08:08:44.858Z  V/SDMSE:Automation:Crawler: crawl():------------12: className=android.widget.TextView, text='677,53 MB used in internal storage', isClickable=false, isEnabled=true, viewIdResourceName=com.android.settings:id/summary, pkgName=com.android.settings, identity=20c7631
```

but the `OnTheFlyLabler` only provided `678 MB` and not the `677,53 MB`.