Core: Check and handle invalid app inventory data on Samsung devices
`QUERY_ALL_PACKAGES` is a permission that is automatically granted, at least according to documentation.
Except, of course, on some Samsung devices running Android 14, because why not do things differently...

This PR prevents SD Maid from crashing in these cases. A specific error is shown.
The state is now detected and treated as a permission error and will show as incomplete-setup.

Fixes #1354