AppCleaner: If a file can't be accessed, continue with the other files
A file that we are trying to delete, may be un-readable in addition to being un-writable.
So a `ReadException` could be thrown too.

While we can't fix the underlying issue that prevents a file from being read or written, we can continue on with the other deletion tasks, and just leave this one for the user to investigate.

Closes #1391