AppCleaner: Fix ACS based cache deletion on HyperOS2 ROMs with updated Security app
The updated "Security" app `com.miui.securitycenter` v10.6.5-250318.1.1(40001065) has changed the UI layout.
It now uses uncommon RadioButtons that are not clickable in an animated bottom sheet dialog. There is no valid `PERFORM_CLICK` target available.
We have to fall back to emulating a manual "tap" gesture and hide the overlay shortly to be able to perform it.
We also have to wait until the bottom sheet dialog animation has settled so we can get the correct coordinates of the element we want to tap.