AppCleaner: Improve ACS based cache deletion performance
The new Security app (`com.miui.securitycenter`) on HyperOS2 poses new challenges.
SD Maid has to wait for the animation of BottomSheetDialog to settle.
The previous approach of iterating through events via the stepper was not powerful enough.
We need to wait until the coordinates of the target Button are no longer changing, only then is the animation done.
Tracking the moving button was not possible in the way the Spec+Stepper logic was setup before.

We need to give specifics specs (i.e. here the HyperOS spec) the ability to wait, block and loop through events on their own.
This makes the stepper tool a bit more dump and moves more of the logic into the spec classes.

Now each spec provides the stepper with a function that gets a stepper context and that function is only called once by the stepper.
The function only returns when its conditions are fulfilled, whether that is done by polling or not is up to each ROM specific spec.

This improves performance for all ROM types as we no longer poll the windowRoot to check all nodes.
Instead we now react to window events, and then check the `windowRoot`.
Less delay based, more event based.