AppCleaner: Fix ACS based cache deletion on HyperOS with latest Security app
In the 2 weeks I'm trying to fix this, Xiaomi updated their Security app again (╯°□°)╯︵ ┻━┻

* First it was a clickable TextView
* Then it was an unclickable RadioButton
* Now it is a TextView again, but it is unclickable too

`v10.6.5-250318.1.1` was still the unclickable RadioButton, some version after that, it is the unclickable TextView.

I've changed the matching to ignore the type and just look for the label.
Less specific, might cause issues in the future, but I'm really tired of this.