AppCleaner: Fix ACS navigating to "Data usage" instead of "Storage"
For ROMs where exact label matching fails, SD Maid will try to find the storage entry based on the storage size which is usually displayed on the entry.
The "On the fly labler" did "contains" matching and in cases where "Data usage" had a summary like "1.80 MB ..." and the size to look for was "80 MB", we got a false positive match and SD Maid got stuck.