AppCleaner: Smarter "Storage" entry finding (fixes false-positives on tablet layouts)
Previously we just went with the first "Storage" entry we found, but sometimes there are more, and the first one is not the one we want.
This is the case on tablets or large screen devices, where the settings UI uses a "double pane" layout.
Here the left pane will also contain a "Storage" entry that takes us to the whole devices storage overview.

Fixes #1720