Setup: Ask for MANAGE_EXTERNAL_STORAGE on AndroidTVs running Android 13+.
Tested this on an emulator. Since Android 13 (API33), AndroidTV also needs to request `MANAGE_EXTERNAL_STORAGE`.
This is a bit later than on phones where we needed that since Android 11 (API30).

Could already be manually granted via settings, but now the setup card takes the user to the right menu.