AppCleaner: Detect if `com.miui.securitycenter` is missing the `GET_USAGE_STATS` permission
No idea how this happens but have seen this in a few reports now.

For some reason, on some devices, the security center app is missing that permission, which causes the `Clear data` button to be disabled, and thus ACS based cache clearing fails.

This change tries to detect that circumstance and shows an error dialog explaining it.