Setup: Check `DocumentsUI` version to determine SAF grantability
The `DocumentsUI` app is part of "Google Play system updates".
A newer version (>= `331120000`) restricts access to `Android/data` and `Android/obb` even on Android 11 & 12 (API 30-32).
This change adds a version check for `DocumentsUI` to correctly reflect its capabilities.

If the version is too new, or if the package information cannot be retrieved, assume access is restricted.
The target SDK version check remains as a fallback.

Fixes https://github.com/d4rken-org/sdmaid-se/discussions/1762