Test: Refine `SegmentCriterium` path matching
`basePath.child("abc", "/")` created an invalid test case as `/` can not be the name of a child.

This commit removes a redundant test case from `SystemCrawlerSieveTest` that was already covered by `SegmentCriteriumTest`.
It also adds new test cases to `SegmentCriteriumTest` to verify correct matching of paths ending with a separator, especially when `allowPartial = true`.